refactor(snippets): move code to util
