+++ src/__tests__/snippets/snippet.test.ts
@@ -7,9 +7,9 @@ import { URI } from 'vscode-uri'
 import events from '../../events'
 import { addPythonTryCatch, evalCode, executePythonCode, getInitialPythonCode, getVariablesCode, hasPython } from '../../snippets/eval'
 import { Placeholder, SnippetParser, Text, TextmateSnippet } from '../../snippets/parser'
-import { CocSnippet, getNextPlaceholder, getTextAfter, getTextBefore, reduceTextEdit } from '../../snippets/snippet'
+import { CocSnippet, getNextPlaceholder } from '../../snippets/snippet'
 import { SnippetString } from '../../snippets/string'
-import { convertRegex, normalizeSnippetString, shouldFormat, UltiSnippetContext } from '../../snippets/util'
+import { convertRegex, getTextAfter, getTextBefore, normalizeSnippetString, reduceTextEdit, shouldFormat, UltiSnippetContext } from '../../snippets/util'
 import { padZero, parseComments, parseCommentstring, SnippetVariableResolver } from '../../snippets/variableResolve'
 import { UltiSnippetOption } from '../../types'
 import { getEnd } from '../../util/position'
